Moved eddy workflow and functions to util for organization
